Updated CRUD operations by protecting the routes with access tokens

+++ controllers/contact-controllers.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc Get a particular contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getAContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -25,7 +25,7 @@ const getAContact = asyncHandler(async (req, res) => {
 
 //@desc Create a contact
 //@route POST /api/contacts/:id
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is:", req.body);
@@ -38,19 +38,25 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json(contact);
 });
 
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("No contact found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(" user has no authorization to update other contacts");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -62,13 +68,18 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete a contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("No contact found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(" user has no authorization to update other contacts");
   }
 
   await Contact.deleteOne({ _id: req.params.id });
